sm7350-common: Build frameworks sensor service hal
* Fixed

04-02 06:01:05.212  1544  1544 W HidlServiceManagement: Waited one
second for android.frameworks.sensorservice@1.0::ISensorManager/default
04-02 06:01:05.213   627   627 I hwservicemanager: Since
android.frameworks.sensorservice@1.0::ISensorManager/default is not
registered, trying to start it as a lazy HAL.
04-02 06:01:05.213  1544  1544 I HidlServiceManagement: getService:
Trying again for
android.frameworks.sensorservice@1.0::ISensorManager/default...
04-02 06:01:05.214   627  4117 W libc    : Unable to set property
"ctl.interface_start" to
"android.frameworks.sensorservice@1.0::ISensorManager/default": error
code: 0x20
04-02 06:01:05.214   627  4117 I hwservicemanager: Tried to start
android.frameworks.sensorservice@1.0::ISensorManager/default as a lazy
service, but was unable to. Usually this happens when a service is not
installed, but if the service is intended to be used as a lazy service,
then it may be configured incorrectly.
04-02 06:01:05.213     1     1 E init    : Control message: Could not
find 'android.frameworks.sensorservice@1.0::ISensorManager/default' for
ctl.interface_start from pid: 627 (/system/bin/hwservicemanager)

Signed-off-by: EndCredits <alicization.han@gmail.com>